Swap precisions

+++ nemo/collections/nlp/modules/common/megatron/transformer.py
@@ -687,11 +687,11 @@ class ParallelTransformerLayer(ParallelTransformerLayer_):
         )
 
         if precision == 'bf16':
-            self.dtype = torch.float32
+            self.dtype = torch.bfloat16
         elif int(precision) == 16:
             self.dtype = torch.float16
         elif int(precision) == 32:
-            self.dtype = torch.bfloat16
+            self.dtype = torch.float32
         else:
             raise ValueError
